Move tests into tests/
